% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classification_celllines.R,
%   R/test_classification_generic.R
\name{cellline_test}
\alias{cellline_test}
\title{DrugPrediction TRAINING}
\usage{
cellline_test(Xtrain, X_test, Ytrain, state)

cellline_test(Xtrain, X_test, Ytrain, state)
}
\arguments{
\item{X_test}{(matrix) matrix}

\item{state}{(vector)}

\item{path_to_bccl}{(string): Path to BCCL normCPMRUV_BCCL.tsv}

\item{path_to_test}{(string): Path to Test matrix}

\item{path_to_combined}{(string): path to combined training and test matrix}

\item{train_idx}{(numeric): numeric list of index of training data sets in path_to_combined}

\item{cell_line_response}{(string): cell_line_response_threshold_0.50_large_and_small_screen.RData}

\item{path_to_targetid}{(string): path to target_id.txt. Used to filter protein_coding genes only.}

\item{path_to_trusight}{(string): path to gene list Trusight.csv. Used to filter genes associated with cancer}

\item{X_train}{(matrix) matrix of expression}

\item{Y_train}{(matrix) matrix of response binary (0,1)}
}
\value{
smmart_trained_machine_learning_model.RData

prediction
}
\description{
#' Testing function for classification on cell lines with prediction and expression
}
\examples{
Y_predicted = cellline_test(Xtrain, X_test, Ytrain, state) 
Y_predicted = cellline_test(Xtrain, X_test, Ytrain, state) 
}
\author{
Janice Patterson

Janice Patterson
}
