% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_classificiation_celllines.R
\name{cellline_train}
\alias{cellline_train}
\title{Training functions for classification on cell lines with prediction and expression}
\usage{
cellline_train(X_train, Y_train)
}
\arguments{
\item{X_train}{(matrix) matrix of expression}

\item{Y_train}{(matrix) matrix of response binary (0,1)}
}
\value{
state
}
\description{
Training functions for classification on cell lines with prediction and expression
}
\examples{
state = cellline_train(X_train, Y_train)
}
\author{
Janice Patterson
}
