% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/kbmtl_semisupervised_classification_variational_test.R
\name{kbmtl_semisupervised_classification_variational_test}
\alias{kbmtl_semisupervised_classification_variational_test}
\title{kbmtl_semisupervised_classification_variational_test
testing procedure for classification}
\usage{
kbmtl_semisupervised_classification_variational_test(K, state)
}
\arguments{
\item{K}{(matrix)}

\item{state}{(vector)}
}
\value{
vector of matrix and vector
}
\description{
kbmtl_semisupervised_classification_variational_test
testing procedure for classification
}
\examples{
kbmtl_semisupervised_classification_variational_test(K_test,state)
}
\author{
Mehmet Gonen
}
