% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/kbmtl_semisupervised_regression_variational_train.R
\name{kbmtl_semisupervised_regression_variational_train}
\alias{kbmtl_semisupervised_regression_variational_train}
\title{kbmtl_semisupervised_regression_variational_train}
\usage{
kbmtl_semisupervised_regression_variational_train(K, Y, parameters)
}
\arguments{
\item{K}{(matrix) matrix}

\item{Y}{(matrix) matrix}

\item{parameters}{(vector)}
}
\value{
state
}
\description{
kbmtl_semisupervised_regression_variational_train
}
\author{
Mehmet Gonen
}
