% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_classification_generic.R
\name{train_classification}
\alias{train_classification}
\title{Train KBMTL classification model generic}
\usage{
train_classification(path_to_combined, train_idx, cell_line_response,
  path_to_targetid, path_to_trusight, normalization_type = "CPMRUV",
  classification_type = "trusight")
}
\arguments{
\item{Xtrain}{(matrix): training data. Featrues are columns and samples are rows}

\item{Ytrain}{(matrix): response}
}
\value{
smmart_trained_machine_learning_model.RData
}
\description{
Train KBMTL classification model generic
}
\examples{

}
