% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_classification_celllines.R
\name{train_classification_celllines}
\alias{train_classification_celllines}
\title{Train KBMTL classifer on Breast Cancer celllines}
\usage{
train_classification_celllines(combined_matrix, train_idx,
  cell_line_response = "cell_line_response_threshold_0.50_large_and_small_screen.RData",
  targetid = "target_id.txt", genelist = "Trusight_genes.csv")
}
\arguments{
\item{combined_matrix}{(string): Tab separated table containing training and test data. Genes are in the rows and samples are in the columns.}

\item{train_idx}{(numeric): Numeric comma separated list of index of the columns that contain the training data from the table in path_to_combined. Index 0 is the gene names. Ranges are colon spearated. eg. 1,2,4:10}

\item{cell_line_response}{(string): cell_line_response_threshold_0.50_large_and_small_screen.RData. Cell line are rows and drugs are columns.}

\item{targetid}{(string): path to target_id.txt. Converts to HUGO identifiers and filters to protein coding genes.}

\item{genelist}{(string): path to gene list Trusight.csv. Used to filter genes associated with cancer. Make sure gene identifiers match.}
}
\value{
smmart_trained_machine_learning_model.RData
}
\description{
Train KBMTL classifer on Breast Cancer celllines
}
\author{
Janice Patterson
}
