% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_classification_celllines.R
\name{cellline_test}
\alias{cellline_test}
\title{Testing function for classification on cell lines with prediction and expression}
\usage{
cellline_test(Xtrain, X_test, Ytrain, state)
}
\arguments{
\item{X_test}{(matrix) matrix}

\item{state}{(vector)}

\item{X_train}{(matrix) matrix of expression}

\item{Y_train}{(matrix) matrix of response binary (0,1)}
}
\value{
prediction
}
\description{
Testing function for classification on cell lines with prediction and expression
}
\examples{
Y_predicted = cellline_test(Xtrain, X_test, Ytrain, state) 
}
\author{
Janice Patterson
}
