% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_classification.R
\name{train_classification}
\alias{train_classification}
\title{Train KBMTL classification model generic}
\usage{
train_classification(Xtrain, Ytrain)
}
\arguments{
\item{Xtrain}{(matrix): training data. Featrues are columns and samples are rows}

\item{Ytrain}{(matrix): response}
}
\value{
smmart_trained_machine_learning_model.RData
}
\description{
Train KBMTL classification model generic
}
\examples{

}
